requirements.txt
pandas>=2.0.0
fastmcp>=0.1.0
pydantic>=2.0.0
typing-extensions>=4.0.0
pytest>=7.0.0
requests>=2.25.0 
